reorg or the dir and namespace tree for Matrics/Trends
