+++ app/Nova/BloodPressureReading.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace App\Nova;
 
-use App\Nova\Metrics\BpmReading;
-use App\Nova\Metrics\DiastolicReading;
-use App\Nova\Metrics\SystolicReading;
+use App\Nova\Metrics\Trends\Readings\Types\BpmReading;
+use App\Nova\Metrics\Trends\Readings\Types\DiastolicReading;
+use App\Nova\Metrics\Trends\Readings\Types\SystolicReading;
 use Illuminate\Http\Request;
 use Laravel\Nova\Fields\DateTime;
 use Laravel\Nova\Fields\ID;
